replace float with double dor coordinates data send to database

+++ database/migrations/2018_03_09_205409_create_points_table.php
@@ -17,8 +17,8 @@ class CreatePointsTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->text('description');
-            $table->float('lattitude');
-            $table->float('longitude');
+            $table->double('lattitude');
+            $table->double('longitude');
             $table->integer('ratingNumVotes');
             $table->integer('safetyNumVotes');
             $table->integer('ratingSumVotes');
